requirements.txt
fastapi
uvicorn
pymupdf
python-docx
scikit-learn
duckduckgo-search
python-multipart
